+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Custom request for client to get doxygen comment (generation of comments) for a line. The request is sent to the server and the server figures-out the the proper comment data.
+- Support for goto declaration and references
 
 ## [1.3.1] - 2025-04-10